requirements.txt
httpx>=0.27.0
markdownify>=0.13.0
ollama>=0.3.0
anthropic>=0.34.0
pyyaml>=6.0.2
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.0.0
